impl. compare and rename for er-macro-transformer
